(PC-12635) refacto(setBirthday): extract BirthdayInformationModal from SetBirthday
